fix(tcp): correctness bugs found in review
- SYN-RECEIVED は acceptable ACK でのみ ESTABLISHED へ昇格 (古い ACK を拒否)
- 受信窓を uint32 化し window scale で 64KB 超バッファを正しく扱う
- IPv4 TotalLength でセグメントを切り詰め、末尾パディング混入を防ぐ
- CLOSED 接続を接続テーブルと Tick 集合から回収し 4-tuple 再利用を可能に
- TIME-WAIT 置換時に新 ISS を旧最大 seq より大きく採る
- TIME-WAIT で FIN 再送に ACK 再送 + 2MSL 再起動で応答
- fast retransmit 後に RTO タイマを再起動し二重送出を防ぐ
- Listener.Close で待機中の Accept を解除
- LISTEN/SYN-SENT の CLOSE を実装

+++ tcp/data.go
@@ -17,8 +17,8 @@ func (c *Conn) SetRcvBuffer(total uint32) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	c.tcb.rcvBuffTotal = total
-	if uint32(c.tcb.rcv.wnd) > total {
-		c.tcb.rcv.wnd = uint16(total)
+	if c.tcb.rcv.wnd > total {
+		c.tcb.rcv.wnd = total
 	}
 }
 
@@ -204,8 +204,8 @@ func (c *Conn) acceptText(h TCPHeader, payload []byte) {
 // consumeRcvWindow は RCV.NXT が n 前進したぶん RCV.WND を減らし右窓端を固定する。
 // 窓未満まで減ったら 0 で飽和させる (右窓端は RCV.NXT を下回らない)。
 func (c *Conn) consumeRcvWindow(n uint32) {
-	if uint32(c.tcb.rcv.wnd) > n {
-		c.tcb.rcv.wnd -= uint16(n)
+	if c.tcb.rcv.wnd > n {
+		c.tcb.rcv.wnd -= n
 	} else {
 		c.tcb.rcv.wnd = 0
 	}
@@ -225,7 +225,7 @@ func (c *Conn) trimToWindow(seq uint32, payload []byte) segFragment {
 		seq = c.tcb.rcv.nxt
 	}
 	// 右トリム: 窓の右端 RCV.NXT+RCV.WND を超えるぶんを捨てる。
-	winEnd := c.tcb.rcv.nxt + uint32(c.tcb.rcv.wnd)
+	winEnd := c.tcb.rcv.nxt + c.tcb.rcv.wnd
 	if SeqGT(end, winEnd) {
 		over := end - winEnd
 		if over >= uint32(len(payload)) {
